pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "tensorflow>=2.14.0",
]
